+++ CHANGELOG.md
@@ -4,9 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 0.1.5 - unreleased
+## 0.1.6 - unreleased
+[#14](https://github.com/BrynCooke/buildstructor/issues/14) Generics ordering bug.
+Generics were not being consistently ordered, which caused issues if there were generics on the impl type and also in a where clause.
+
+## 0.1.5 - 2022-04-11
 ### Added
-[#9](https://github.com/BrynCooke/buildstructor/issues/9) Add `*_new` support
+[#9](https://github.com/BrynCooke/buildstructor/issues/9) Add `*_new` support.
 Any method named `new` or has a suffix `_new` will create a builder.
 Builders methods are named appropriately. e.g. `try_new` -> `try_build`.
 ### Fixed